requirements.txt
llama-index
together
pydantic
python-dotenv